Move backend miscellaneous tets
