+++ packages/backend/src/api/controllers/ForcedTradeOfferController.test.ts
@@ -3,12 +3,12 @@ import { EthereumAddress, Hash256, Timestamp } from '@explorer/types'
 import { expect } from 'earljs'
 import { Wallet } from 'ethers'
 
-import { fakeAccepted, fakeOffer } from '../../../test/fakes'
-import { mock } from '../../../test/mock'
 import { AccountService } from '../../core/AccountService'
 import { ForcedTradeOfferRepository } from '../../peripherals/database/ForcedTradeOfferRepository'
 import { PositionRepository } from '../../peripherals/database/PositionRepository'
 import { UserRegistrationEventRepository } from '../../peripherals/database/UserRegistrationEventRepository'
+import { fakeAccepted, fakeOffer } from '../../test/fakes'
+import { mock } from '../../test/mock'
 import { ForcedTradeOfferController } from './ForcedTradeOfferController'
 import * as tradeMock from './utils/ForcedTradeOfferMockData'
