+++ packages/backend/src/api/controllers/utils/ForcedTradeOfferValidators.test.ts
@@ -3,7 +3,7 @@ import { AssetId, EthereumAddress, StarkKey } from '@explorer/types'
 import { expect } from 'earljs'
 import { Wallet } from 'ethers'
 
-import { fakeBigInt } from '../../../../test/fakes'
+import { fakeBigInt } from '../../../test/fakes'
 import { accepted, addressB, offer } from './ForcedTradeOfferMockData'
 import {
   validateAcceptSignature,
