+++ packages/backend/test/core/PerpetualRollupUpdater.test.ts
@@ -26,8 +26,8 @@ import { PageRepository } from '../../src/peripherals/database/PageRepository'
 import { StateTransitionRecord } from '../../src/peripherals/database/StateTransitionRepository'
 import { StateUpdateRepository } from '../../src/peripherals/database/StateUpdateRepository'
 import type { EthereumClient } from '../../src/peripherals/ethereum/EthereumClient'
+import { mock } from '../../src/test/mock'
 import { Logger } from '../../src/tools/Logger'
-import { mock } from '../mock'
 
 const emptyState: State = {
   positionRoot: PedersenHash.ZERO,
