+++ packages/backend/test/core/collectors/FinalizeExitEventsCollector.test.ts
@@ -14,8 +14,8 @@ import {
   fakeForcedUpdatesVerified,
   fakeInt,
   fakeTimestamp,
-} from '../../fakes'
-import { mock } from '../../mock'
+} from '../../../src/test/fakes'
+import { mock } from '../../../src/test/mock'
 
 const blockRange = new BlockRange([
   {
