+++ packages/backend/test/core/collectors/UserRegistrationCollector.test.ts
@@ -5,7 +5,7 @@ import { UserRegistrationCollector } from '../../../src/core/collectors/UserRegi
 import { BlockRange } from '../../../src/model'
 import { UserRegistrationEventRepository } from '../../../src/peripherals/database/UserRegistrationEventRepository'
 import type { EthereumClient } from '../../../src/peripherals/ethereum/EthereumClient'
-import { mock } from '../../mock'
+import { mock } from '../../../src/test/mock'
 
 const PERPETUAL_ADDRESS = EthereumAddress.fake('deadbeef1234')
